### 2020-02-07 #16 @cboettig
Ten-year Reproduciblity Challenge: Fluctuation Domains
**Original article:** Carl Boettiger, Jonathan Dushoff, Joshua S Weitz (2010). Fluctuation domains in adaptive evolution, Theoretical Population Biology 77 (1) 6-13. doi:10.1016/j.tpb.2009.10.003.

**Scientific domain:**  Ecology & Evolution
**Programming language:**  C, R
