### 2020-02-10 #16 @rougier
@ctb Can you edit this reproduction submission for the Ten Years Reproducibility Challenge (you only need one reviewer) ?
