### 2020-02-13 #16 @rougier
@tpoisot Can you edit this reproduction submission for the Ten Years Reproducibility Challenge (you only need one reviewer) ?
