+++ rescience/manuscript.tex
@@ -203,7 +203,7 @@
 \hypertarget{historical-context}{%
 \subsection{Historical context}\label{historical-context}}
 
-In this analyis I seek to reproduce the numerical results of (Boettiger,
+In this analysis I seek to reproduce the numerical results of (Boettiger,
 Dushoff, and Weitz 2010), which focuses on the evolutionary trajectory
 of species trait under selection, as predicted by the so-called
 ``Canonical Equation'' of adaptive dynamics (Dieckmann and Law 1996).
@@ -249,7 +249,7 @@ personal website of one of my co-authors, with URL
 URL still resolves (my colleague is still at the same institution), and
 does indeed contain a downloadable zip file with a copy of the code. In
 2012, I had uploaded a copy of this code as a supplement to Dryad, but
-both on my CV a the time and on my colleague's website we had recorded
+both on my CV at the time and on my colleague's website we had recorded
 only the Dryad handle URL,
 \url{http://datadryad.org/handle/10255/dryad.37625}, which no longer
 resolves. That handle can be resolved by the WayBackMachine, which
@@ -283,7 +283,7 @@ facilitate reproducibility, stating at the end of the manuscript:
 \begin{quote}
 In the spirit of scientifically reproducible research(Gentleman and
 Lang, 2007; Schwabet al., 2000; Stodden, 2009), we freely provide all
-the source code required to replicate the simulations and figuress hown
+the source code required to replicate the simulations and figures shown
 in the text. Though the numerical simulations are written in C for
 computational efficiency, we provide a user interface and documentation
 by releasing all the code, figures, text, and examples as a software
@@ -316,7 +316,7 @@ wget http://ecotheory.biology.gatech.edu/sites/default/files/adpaper_0.2-8.tar.g
 R CMD INSTALL adpaper_0.2-8.tar.gz
 \end{verbatim}
 
-The same stategy worked on MacOSX (Catalina) with \texttt{gsl} installed
+The same strategy worked on MacOSX (Catalina) with \texttt{gsl} installed
 via Homebrew. I was able to update the original autoconf script by
 recalling that it had been adapted from another package, the
 \texttt{gsl} R package
@@ -335,22 +335,22 @@ pernicious, because versioning of these components is both platform
 dependent and separate from the R package manager process, and thus
 cannot be addressed merely with package version tracking in tools such
 as \texttt{packrat} (Ushey et al. 2018) or \texttt{renv} (Ushey 2019).
-Today, containerized approaches such a Docker (see Boettiger 2015;
+Today, containerized approaches such as Docker (see Boettiger 2015;
 Boettiger and Eddelbuettel 2017) provide a solution to this
 system-dependency issue that was not available a decade ago.
 
 The second challenge was in figuring out just what commands had to be
 run to duplicate the figures. As mentioned above, the GitHub version of
 the source-code did not have the manuscript source, and even the version
-that did have the mansucript source did not write the manuscript itself
+that did have the manuscript source did not write the manuscript itself
 as a dynamic document (e.g.~in Sweave (Xie 2015)). In that version,
 files \texttt{fig1.R} and \texttt{fig2.R} are were included in the same
 subdirectory as the \texttt{pdf} figures referenced by the \texttt{tex}
-version of the manuscript, but \texttt{fig2.R}, but these files
+version of the manuscript, but these files
 contained only the code required to actually generate the figures
 themselves once the result data had been pre-computed and saved as
-\texttt{.txt} files. That result data files had also not been archived
-seperately. Fortunately, re-reading my own documentation, I was able to
+\texttt{.txt} files. The result data files had also not been archived
+separately. Fortunately, re-reading my own documentation, I was able to
 deduce that these text files were generated by running the function
[…]
