+++ rescience/article.Rmd
@@ -7,7 +7,6 @@ output:
     includes:
       before_body: header.tex
       after_body: suffix.tex
-    keep_tex: true
 documentclass: rescience
 bibliography: "bibliography.bib"
 
@@ -20,7 +19,7 @@ knitr::opts_chunk$set(cache=TRUE, warning=FALSE, error=FALSE, message = FALSE)
 
 ## Historical context
 
-In this analsyis I seek to reproduce the numerical results of [@Boettiger2010], which focuses on the evolutionary trajectory of species trait under selection, as predicted by the so-called "Canonical Equation" of adaptive dynamics [@Dieckmann1996].  Theoretical work I present in the paper predicts the emergence of a 'flutation enhancement' regime far from an evolutionary stable point, and suggests that the evolution of traits which start in this distant region would deviate substantially from the expected evolutionary path predicted by the canonical equation.  In particular, the distribution of such trajectories could be bimodal, making the "canonical" path far from the most likely path.  The details of this distribution cannot be solved for analytically however, so in this paper I resorted to an replicate exact stochastic simulations of the evolutionary model using a Gillespie algorithm [@Gillespie].
+In this analysis I seek to reproduce the numerical results of [@Boettiger2010], which focuses on the evolutionary trajectory of species trait under selection, as predicted by the so-called "Canonical Equation" of adaptive dynamics [@Dieckmann1996].  Theoretical work I present in the paper predicts the emergence of a 'fluctuation enhancement' regime far from an evolutionary stable point, and suggests that the evolution of traits which start in this distant region would deviate substantially from the expected evolutionary path predicted by the canonical equation.  In particular, the distribution of such trajectories could be bimodal, making the "canonical" path far from the most likely path.  The details of this distribution cannot be solved for analytically however, so in this paper I resorted to an replicate exact stochastic simulations of the evolutionary model using a Gillespie algorithm [@Gillespie].
 
 I implemented Gillespie algorithm for the evolutionary model in C code, and wrapped the C code in an R package structure to make it easier to run and easier to plot results.  The C code depended on one standard C library, the GNU Scientific Library, which has evolved considerably (with one major version change) since that time.  The original source code was 'published' (initially only on a colleague's website, as stated in the paper, though later copies were submitted to formal archives, as discussed below).  That original copy was distributed under a GPL v3 licence, though I later replaced that with a CC0 license when submitting the version to a data archive (Dryad) which required public domain declarations.
 
@@ -31,7 +30,7 @@ I have changed laptops and the organizational schemes I use for my documents sev
 
 The published paper itself states that the code is available at the personal website of one of my co-authors, with URL <http://ecotheory.biology.gatech.edu/downloads/>.  Fortunately, this URL still resolves (my colleague is still at the same institution), and does indeed contain a downloadable zip file with a copy of the code. In 2012, I had uploaded a copy of this code as a supplement to Dryad, but both on my CV at the time and on my colleague's website we had recorded only the Dryad handle URL, <http://datadryad.org/handle/10255/dryad.37625>, which no longer resolves.  That handle can be resolved by the WayBackMachine, which doesn't contain the code archive but does reveal the DOI, <http://dx.doi.org/10.5061/dryad.j8n0p7vc>, which successfully resolves to the Dryad copy.  The code can also be found by searching the Dryad web interface for my surname (though it is listed as "data" and not "software."  In 2017, I had also imported the GitHub copy into the Zenodo data repository where it was again assigned a DOI and is identified as software in the metadata.  I note that the package was never submitted to the Central R Archive Network, CRAN.
 
[…]
