### 2020-02-21 #16 @rougier
Gentle reminder: @tpoisot Can you edit this reproduction submission for the Ten Years Reproducibility Challenge (you only need one reviewer) ?
