### 2020-02-24 #16 @ha0ye
The languages and topic do seem in my wheelhouse. I won't have time until the 2nd week of March (at the earliest), though.

For transparency, I also interact with @cboettig in a variety of places, but we haven't and aren't working on anything together. He did invite me to be on a panel in 2018. (Not sure if any of that matters for COI purposes.)
