### 2020-02-28 #16 @rougier
I think we do not consider this as a COI.
