### 2020-04-09 #16 @ha0ye
## Review of 'Reproducing "Fluctuation Domains"'

In reviewing this replication of (Boettiger, Carl, Jonathan Dushoff, and Joshua S. Weitz. 2010. “Fluctuation Domains in Adaptive Evolution.”), I follow the 4-point criteria listed on https://rescience.github.io/edit/

1. The actual replication of the research. The reviewer must evaluate the authors’ claims about a successful replication, applying the standards of the field.

The results from the original paper consist mainly of a numerical simulation (Figure 2 in the original paper). This has been successfully reproduced with the updated code (Figures 4-6 of the replication).

Figure 1 from the original paper has also been reproduced successfully (Figures 1-3 of the replication).

2. Reproducibility of the replication. The reviewers must be able to run the proposed implementation on their computer, and obtain the same results as the authors with the limits of the state of the art of the field.

No issues here. :+1:

3. Clarity of the code and the instructions for running it. Uncommented or obfuscated code is as bad as no code at all.

Both the code and the replication are presented in the form of an R package compendium, with the changes for the replication within a branch of the GitHub repo for the code accompanying the original paper, and a `rescience` folder that contains files associated with the replication paper.

The package functions appear to perform the dual task of doing computations and then plotting the results. This could be better documented (with function names that are more descriptive) and/or refactored (so that computations and plotting are done separately). The latter adds flexibility to save intermediate results and/or subsequent analyses.

Finally, it would be useful to document the replication status and changes to the code within the top-level README. Some basic installation instructions could be added (particularly the note about installing GSL prior to the usual instructions for installing an R package).

4. Clarity and completeness of the accompanying article, in which the authors should clearly state why they think they have replicated the paper (same figures, same graphics, same behavior, etc.) and explain any obstacles they have encountered during the replication work.

The article was straightforward to follow, with detailed discussion of issues around dependency versioning and changing standards. One suggestion I have is to set the chunk options so that the current Figures 1-3 are panels within a single figure, like in the original paper. (Currently, they are separate figures, with duplicated captions.)

These are the chunk settings that I used to re-knit the document:
```
