### 2020-04-21 #16 @cboettig
Thanks @ha0ye, these are great suggestions.

@rougier @tpoisot Just checking in if anything is needed from me here yet?
