### 2020-05-03 #16 @rougier
@cboettig You can edit your submission following @ha0ye suggestions and then @tpoisot will (probably) accept the article. Then there will be some work for the actual publication (mostly editing your metadata file).
