### 2020-05-16 f3ddb05
Address reviewer suggestions
Adds the chunk options and README suggested by @ha0ye.

While I agree that it would be cleaner to seperate out the plotting code from the code calling the C routines, I think this might in fact lead to more confusion, since the casual observer might wrongly conclude such more involved changes were necessary for the successful reproduction of the results.

Other than this, I believe this addresses all of the reviewer's suggestions.  Thanks again for taking the time to examine this piece and run the code!
