### 2020-05-18 #16 @cboettig
Thanks @rougier.  @tpoisot I've made the edits in https://github.com/cboettig/fluctuationDomains/pull/3.
