### 2020-06-02 #16 @cboettig
@rougier @tpoisot :point_up:
