### 2020-09-09 #16 @cboettig
@rougier Anything we can do here?  COVID and all the other crises might be interfering with @tpoisot 's ability to look at this.  I think the changes requested and implemented above are pretty straight forward, maybe another independent editor can assess?
