### 2020-09-10 #16 @rougier
You're right, I'll handle the submission from now on. So congratulation, your article has been accepted!
Can you give me the link to your article latest repo?
