### 2020-09-11 #16 @cboettig
@rougier I've merged the above-mentioned PRs, so everything is now in https://github.com/cboettig/fluctuationDomains/tree/master/rescience, with the pdf at https://github.com/cboettig/fluctuationDomains/blob/master/rescience/article.pdf, and article source here: https://github.com/cboettig/fluctuationDomains/blob/master/rescience/article.Rmd

(I can add the combined `.tex` file that will be generated from the `.Rmd` if you want -- for us Rmd folks tex is just an intermediate object between .Rmd source and .pdf output :-) )
