### 2020-09-15 #16 @rougier
I tried compiling from the .Rmd but I had some problems. If you can add the article.tex, that would be great (I need to compile it several times to test the publication process + update metadata)
