### 2020-09-15 #16 @rougier
BTW, I got:
```
Error in library(fluctuationDomains) :
  there is no package called 'fluctuationDomains'
```
when inside the rescience directory. I'm not a R user but I imagine I've to tell R I'm using a local package somehow.
