### 2020-09-15 e547c62
include intermediate tex add installation to Makefile
