### 2020-09-15 #16 @cboettig
Thanks @rougier !

- `tex` version is now here: https://github.com/cboettig/fluctuationDomains/blob/master/rescience/article.tex
- I added an `install` target to the Makefile, which will install `fluctuationDomains` locally.  (you'll just need a copy of libgsl-dev on your machine).  Typically we R folks install directly from GitHub by running `remotes::install_github("cboettig/fluctuationDomains")` in R, which has the added benefit of automatically installing any R package dependencies, but in this case there aren't any :-).
