### 2020-09-18 #16 @rougier
Perfect, it worked. Could you save your github repository at https://www.softwareheritage.org/save-and-reference-research-software/ and give me back the swh id you'll get. Also, just before the conclusion, you have a mising DOI that I think should be a reference to the github repository. It would be better to use the swh id that will in turn points to the github repo. If you update your article.tex, I can merged it in my local fork.
