### 2020-09-23 #16 @rougier
@cboettig ping
