### 2020-09-23 #16 @cboettig
Thanks, queued it on software heritage and forgot to check back next day (queue seems to take a while! though I noticed anyone can queue any public repo, you don't have to be the owner).  It's archived now: https://archive.softwareheritage.org/browse/origin/directory/?origin_url=https://github.com/cboettig/fluctuationDomains&timestamp=2020-09-19T00:48:22.704238%2B00:00

Um, which software heritage ID do you recommend I put into the article.tex in place of the DOI?  This one? https://archive.softwareheritage.org/browse/revision/e547c62fd9ae8361d601302200bbc90b26299007/?origin_url=https://github.com/cboettig/fluctuationDomains&timestamp=2020-09-19T00:48:22Z or is there something more compact?
