### 2020-09-24 #16 @rougier
Didn't you get a `SWH...` id at some point? It think the link you get is for a specifc revision(which make sense) but I think there should be an initial SWD id. See for example http://rescience.github.io/bibliography/DiCosmo_2020.html by @rdicosmo
