### 2020-09-24 #16 @cboettig
Thanks @rougier, but even with the DiCosmo example I'm not sure what you are referring to as "the Software Heritage Identifier".  I see a DOI there that belongs to the Zenodo record, and I'd be happy to provide that.  The link I see that is specific to software heritage is https://archive.softwareheritage.org/swh:1:rev:2db189928c94d62a3b4757b3eec68f0a4d4113f0;origin=https://gitorious.org/ocamlp3l/ocamlp3l_cvs.git/.

I'm a huge fan of software heritage.  One thing I think is absolutely brilliant about their approach is that they use content-based identifiers, instead of issuing arbitrary unique strings that are specific to software heritage.  You'll not that the sha1 hashes showing up in my link and in the DiCosmo link are literally the hashes of the most recent commit.  Software heritage "understands" and can "resolve" these IDs, so my most recent git commit appended with the right prefix `swh:1:rev:e547c62fd9ae8361d601302200bbc90b26299007` could be considered my ID, but note that this string doesn't "come from" software heritage, it existed before the Software Heritage archive was created, and before I even pushed that commit to git. Now that the repo is in the SWH archive, SWH can of course resolve that commit hash, or any other commit hash from the repo.

SWH can also resolve individual objects by their SHA-256 sum; e.g. you can literally `sha256sum article.pdf` in my repo, and then [query SWH](https://archive.softwareheritage.org/api/1/content/sha256:9bdf2439e0fd03887fc6c0dda0f6c3d867e6c58869ed1ca4e5d534c4846da052/raw/) by that to get just those bits-and-bytes instead of the whole repo.  Note this is completely independent of git's hashes.

All this is great and I wish other archives like Zenodo would also allow these content-based queries, but I'm not quite sure what to put in the paper.  Readers probably don't know how to resolve a commit hash from software heritage, unless we embed the full link to their API (in which case they probably won't know what the URL signifies, it just looks like a long URL).  I think readers expect a DOI and understand what that means; it seems even Prof DiCosmo has provided a Zenodo DOI in addition to the SWH URL.  Advice on this?  I think you're blazing a trail here.
