### 2020-10-01 #16 @rougier
I think we can use the `swh:1:rev:e547c62fd9ae8361d601302200bbc90b26299007` using the address https://archive.softwareheritage.org/swh:1:rev:e547c62fd9ae8361d601302200bbc90b26299007 and the ReScience template should make this cickable. Let me include it in your article and you tell me if it look ok (I'll add the Zenodo DOI as well).
