### 2020-10-01 #16 @rougier
It's up at Zenodo **sandbox** server: https://sandbox.zenodo.org/record/676290 (it will be deleted by tomorrow)
