### 2020-10-01 #16 @rougier
@cboettig can you check the PDF for any error? I modified a bit your article.tex to correct layout & fonts.
