### 2020-10-06 #16 @cboettig
@rougier apologies missed my github notifications here.  this all sounds good, let me know if you want me to check it over again.
