### 2020-10-06 #16 @cboettig
Also I did request a DOI from zenodo for the `rescience-accepted` tagged version: https://doi.org/10.5281/zenodo.4067657 if you want to use that.
